requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.21.6
Pillow>=9.0.0
PyYAML>=5.3.1
tqdm>=4.64.0

# YOLOv5 specific
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
scipy>=1.4.1

# Model export and inference
onnx>=1.12.0
onnxruntime>=1.15.0
onnxsim>=0.4.13

# Utilities
tensorboard>=2.4.1
thop>=0.1.1

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
